+++ .storybook/StoryWrapper.tsx
@@ -1,0 +1,11 @@
+import React from 'react';
+
+import './StoryWrapper.css';
+
+export function StoryWrapper(props: React.PropsWithChildren<any>): JSX.Element {
+  return (
+    <React.Suspense fallback={''}>
+      <div>{props.children}</div>
+    </React.Suspense>
+  );
+}