+++ packages/codeblock-react/src/Codeblock.async.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 
-import { setAutoloadPath } from '@codeblock/core';
+import { setAutoload } from '@codeblock/core/lib/http';
 import { ProviderConfig } from '@codeblock/core/types';
 import asyncLanguageProvider from '@codeblock/languages/lib/async';
 import asyncThemeProvider from '@codeblock/themes/lib/async';
@@ -8,16 +8,13 @@ import asyncThemeProvider from '@codeblock/themes/lib/async';
 import { Codeblock } from './Codeblock';
 import { CodeblockProps } from './types';
 
-const asyncProviderConfig: ProviderConfig = {
+export const asyncProviders: ProviderConfig = {
   themes: asyncThemeProvider,
   languages: asyncLanguageProvider
 };
 
 export const CodeblockAsync = (props: CodeblockProps) => {
-  React.useEffect(() => {
-    setAutoloadPath(null);
-  }, []);
-  return <Codeblock providers={asyncProviderConfig} {...props} />;
+  return <Codeblock providers={asyncProviders} {...props} />;
 };
 
 export default CodeblockAsync;