+++ packages/codeblock-react/stories/Codeblock.cdn.stories.tsx
@@ -1,22 +1,22 @@
 import raw from 'raw.macro';
 import React from 'react';
 import { withKnobs } from '@storybook/addon-knobs';
+import { CodeblockStoryExample } from './storybook-helpers';
+import { StoryWrapper } from '../../../.storybook/StoryWrapper';
 
-import { CodeblockStoryExample } from './index.stories';
-
-export default { title: 'Codeblock.http', decorators: [withKnobs] };
+export default { title: 'Codeblock.react', decorators: [withKnobs] };
 
 const CodeblockCDN = React.lazy(() => import('../src/Codeblock.cdn'));
 
-export const CDN = () => {
+export const With_CDN_Provider = () => {
   return (
-    <div>
+    <StoryWrapper>
       <h3>HTTP - CDN</h3>
       <p>Same as the HTTP example, but resources are loaded from a CDN.</p>
       <CodeblockStoryExample
         codeblockRenderer={CodeblockCDN}
-        initialContent={raw('../src/Codeblock.http.tsx')}
+        children={raw('../src/Codeblock.http.tsx')}
       />
-    </div>
+    </StoryWrapper>
   );
 };