+++ packages/codeblock-themes/src/cdn.ts
@@ -1,8 +1,8 @@
 import {
-  HTTP_AUTOLOAD_PATH,
+  CDN_AUTOLOAD_PATH,
   createHttpThemeProvider
 } from '@codeblock/core/lib/http';
 
 import { loadPrismTheme } from './utils/load-prism-theme';
 
-export default createHttpThemeProvider(HTTP_AUTOLOAD_PATH, loadPrismTheme);
+export default createHttpThemeProvider(CDN_AUTOLOAD_PATH, loadPrismTheme);