+++ packages/codeblock-core/CHANGELOG.md
@@ -3,6 +3,108 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.0.6](https://github.com/codeblockjs/codeblock/compare/@codeblock/core@2.0.6-alpha.0...@codeblock/core@2.0.6) (2020-02-14)
+
+**Note:** Version bump only for package @codeblock/core
+
+
+
+
+
+## [2.0.6-alpha.0](https://github.com/codeblockjs/codeblock/compare/@codeblock/core@2.0.5...@codeblock/core@2.0.6-alpha.0) (2020-02-14)
+
+
+### Features
+
+* add usePrismAutoload option ([c143c82](https://github.com/codeblockjs/codeblock/commit/c143c822c8dd330c34b18bad35b138a2a161e81f))
+
+
+
+
+
+## [2.0.5](https://github.com/codeblockjs/codeblock/compare/@codeblock/core@2.0.4...@codeblock/core@2.0.5) (2020-02-13)
+
+
+### Bug Fixes
+
+* do not set invalid autoload url ([09003bc](https://github.com/codeblockjs/codeblock/commit/09003bcb3f722bcf4f5848a44a053b02a4627720))
+
+
+
+
+
+## [2.0.4](https://github.com/codeblockjs/codeblock/compare/@codeblock/core@2.0.3...@codeblock/core@2.0.4) (2020-02-13)
+
+
+### Bug Fixes
+
+* allow unsetting autoload path ([b77625d](https://github.com/codeblockjs/codeblock/commit/b77625dbae103be10b55fc282c98848c5c620948))
+
+
+
+
+
+## [2.0.3](https://github.com/codeblockjs/codeblock/compare/@codeblock/core@2.0.2...@codeblock/core@2.0.3) (2020-02-13)
+
+**Note:** Version bump only for package @codeblock/core
+
+
+
+
+
+## [2.0.2](https://github.com/codeblockjs/codeblock/compare/@codeblock/core@2.0.1...@codeblock/core@2.0.2) (2020-02-13)
+
+**Note:** Version bump only for package @codeblock/core
+
+
+
+
+
+## [2.0.1](https://github.com/codeblockjs/codeblock/compare/@codeblock/core@2.0.0...@codeblock/core@2.0.1) (2020-02-12)
+
+
+### Bug Fixes
+
+* detect languages in nested children ([b1c3cc9](https://github.com/codeblockjs/codeblock/commit/b1c3cc989c4f5fd81b1d98e8d0b7966d02b87134))
+
+
+
+
+
+# 2.0.0 (2020-02-12)
+
+
+### Features
+
+* expose async and empty provider configs ([eecb909](https://github.com/codeblockjs/codeblock/commit/eecb909e79dc2d199a0ddd95a306824f3af1894a))
+
+
+
+## 1.0.1 (2020-02-12)
+
+
+
+# 1.0.0 (2020-02-12)
+
+
+### Features
+
+* add http and cdn providers ([b13b907](https://github.com/codeblockjs/codeblock/commit/b13b9076ca2a0ddf637bc2e102da6490f6b66a2e))
+* add http language provider ([72bf8c2](https://github.com/codeblockjs/codeblock/commit/72bf8c2a2db8557a65fc39e009cd420561b507a2))
+
+
+### BREAKING CHANGES
+
+* drop @codeblock/styled
+
+
+
+## 0.0.2 (2020-02-10)
+
+
+
+
+
 ## [1.0.1](https://github.com/codeblockjs/codeblock/compare/v1.0.0...v1.0.1) (2020-02-12)
 
 **Note:** Version bump only for package @codeblock/core