+++ packages/codeblock-languages/CHANGELOG.md
@@ -3,6 +3,97 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.0.7](https://github.com/codeblockjs/codeblock/compare/@codeblock/languages@2.0.7-alpha.0...@codeblock/languages@2.0.7) (2020-02-14)
+
+**Note:** Version bump only for package @codeblock/languages
+
+
+
+
+
+## [2.0.7-alpha.0](https://github.com/codeblockjs/codeblock/compare/@codeblock/languages@2.0.6...@codeblock/languages@2.0.7-alpha.0) (2020-02-14)
+
+
+### Features
+
+* add cdn and http providers ([99ac9fb](https://github.com/codeblockjs/codeblock/commit/99ac9fb57c3cc2dbf0fd5540e2502967fdd239ae))
+
+
+
+
+
+## [2.0.6](https://github.com/codeblockjs/codeblock/compare/@codeblock/languages@2.0.5...@codeblock/languages@2.0.6) (2020-02-13)
+
+**Note:** Version bump only for package @codeblock/languages
+
+
+
+
+
+## [2.0.5](https://github.com/codeblockjs/codeblock/compare/@codeblock/languages@2.0.4...@codeblock/languages@2.0.5) (2020-02-13)
+
+**Note:** Version bump only for package @codeblock/languages
+
+
+
+
+
+## [2.0.4](https://github.com/codeblockjs/codeblock/compare/@codeblock/languages@2.0.3...@codeblock/languages@2.0.4) (2020-02-13)
+
+**Note:** Version bump only for package @codeblock/languages
+
+
+
+
+
+## [2.0.3](https://github.com/codeblockjs/codeblock/compare/@codeblock/languages@2.0.2...@codeblock/languages@2.0.3) (2020-02-13)
+
+**Note:** Version bump only for package @codeblock/languages
+
+
+
+
+
+## [2.0.2](https://github.com/codeblockjs/codeblock/compare/@codeblock/languages@2.0.1...@codeblock/languages@2.0.2) (2020-02-13)
+
+**Note:** Version bump only for package @codeblock/languages
+
+
+
+
+
+## [2.0.1](https://github.com/codeblockjs/codeblock/compare/@codeblock/languages@2.0.0...@codeblock/languages@2.0.1) (2020-02-12)
+
+**Note:** Version bump only for package @codeblock/languages
+
+
+
+
+
+# 2.0.0 (2020-02-12)
+
+
+
+## 1.0.1 (2020-02-12)
+
+
+
+# 1.0.0 (2020-02-12)
+
+
+### Features
+
+* add http and cdn providers ([b13b907](https://github.com/codeblockjs/codeblock/commit/b13b9076ca2a0ddf637bc2e102da6490f6b66a2e))
+
+
+### BREAKING CHANGES
+
+* drop @codeblock/styled
+
+
+
+
+
 ## [1.0.1](https://github.com/codeblockjs/codeblock/compare/v1.0.0...v1.0.1) (2020-02-12)
 
 **Note:** Version bump only for package @codeblock/languages