+++ packages/codeblock-react/CHANGELOG.md
@@ -3,6 +3,238 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [2.3.0](https://github.com/codeblockjs/codeblock/compare/@codeblock/react@2.2.1...@codeblock/react@2.3.0) (2020-02-14)
+
+
+### Features
+
+* add `src` prop for auto-loading external content ([8684d4f](https://github.com/codeblockjs/codeblock/commit/8684d4f616666c7799f6734d3ab23c5651b10908))
+
+
+
+
+
+## [2.2.1](https://github.com/codeblockjs/codeblock/compare/@codeblock/react@2.2.1-alpha.3...@codeblock/react@2.2.1) (2020-02-14)
+
+
+### Bug Fixes
+
+* pass theme to hook ([3e4f924](https://github.com/codeblockjs/codeblock/commit/3e4f9242de3f0bc828093b7d73a1b5f84f44f278))
+
+
+
+
+
+## [2.2.1-alpha.3](https://github.com/codeblockjs/codeblock/compare/@codeblock/react@2.2.1-alpha.2...@codeblock/react@2.2.1-alpha.3) (2020-02-14)
+
+**Note:** Version bump only for package @codeblock/react
+
+
+
+
+
+## [2.2.1-alpha.2](https://github.com/codeblockjs/codeblock/compare/@codeblock/react@2.2.1-alpha.1...@codeblock/react@2.2.1-alpha.2) (2020-02-14)
+
+
+### Bug Fixes
+
+* update theme classname despite of missing providers ([9c7f0de](https://github.com/codeblockjs/codeblock/commit/9c7f0de70e341f91b1ef80257e5ab50591ac4ac8))
+
+
+
+
+
+## [2.2.1-alpha.1](https://github.com/codeblockjs/codeblock/compare/@codeblock/react@2.2.1-alpha.0...@codeblock/react@2.2.1-alpha.1) (2020-02-14)
+
+**Note:** Version bump only for package @codeblock/react
+
+
+
+
+
+## [2.2.1-alpha.0](https://github.com/codeblockjs/codeblock/compare/@codeblock/react@2.2.0...@codeblock/react@2.2.1-alpha.0) (2020-02-14)
+
+
+### Features
+
+* use new core ([aa9f36b](https://github.com/codeblockjs/codeblock/commit/aa9f36bd38c8cb980aebf4adeb4fb472e1300747))
+
+
+
+
+
+# [2.2.0](https://github.com/codeblockjs/codeblock/compare/@codeblock/react@2.1.8...@codeblock/react@2.2.0) (2020-02-13)
+
+
+### Features
+
+* add useHTTPProvider hook ([f42e896](https://github.com/codeblockjs/codeblock/commit/f42e896307dd9caa583472d7e1339a0c81cad1c2))
+
+
+
+
+
+## [2.1.8](https://github.com/codeblockjs/codeblock/compare/@codeblock/react@2.1.7...@codeblock/react@2.1.8) (2020-02-13)
+
+**Note:** Version bump only for package @codeblock/react
+
+
+
+
+
+## [2.1.7](https://github.com/codeblockjs/codeblock/compare/@codeblock/react@2.1.6...@codeblock/react@2.1.7) (2020-02-13)
+
+**Note:** Version bump only for package @codeblock/react
+
+
+
+
+
+## [2.1.6](https://github.com/codeblockjs/codeblock/compare/@codeblock/react@2.1.5...@codeblock/react@2.1.6) (2020-02-13)
+
+**Note:** Version bump only for package @codeblock/react
+
+
+
+
+
+## [2.1.5](https://github.com/codeblockjs/codeblock/compare/@codeblock/react@2.1.4...@codeblock/react@2.1.5) (2020-02-13)
+
+**Note:** Version bump only for package @codeblock/react
+
+
+
+
+
+## [2.1.4](https://github.com/codeblockjs/codeblock/compare/@codeblock/react@2.1.3...@codeblock/react@2.1.4) (2020-02-13)
+
+**Note:** Version bump only for package @codeblock/react
+
+
+
+
+
+## [2.1.3](https://github.com/codeblockjs/codeblock/compare/@codeblock/react@2.1.2...@codeblock/react@2.1.3) (2020-02-13)
+
+**Note:** Version bump only for package @codeblock/react
+
+
+
+
+
+## [2.1.2](https://github.com/codeblockjs/codeblock/compare/@codeblock/react@2.1.1...@codeblock/react@2.1.2) (2020-02-12)
+
+
+### Bug Fixes
+
+* use default theme (okaidia) ([ae66873](https://github.com/codeblockjs/codeblock/commit/ae66873d3f763a95f3d26a4b676c96c8d8f2adc9))
+
+
+
+
+
+## [2.1.1](https://github.com/codeblockjs/codeblock/compare/@codeblock/react@2.1.0...@codeblock/react@2.1.1) (2020-02-12)
+
+
+### Bug Fixes
+
+* add default export ([5b6eb00](https://github.com/codeblockjs/codeblock/commit/5b6eb00e318fbce8984d810aab7ba4c7d1ed908e))
+
+
+
+
+
+# [2.1.0](https://github.com/codeblockjs/codeblock/compare/@codeblock/react@2.0.0...@codeblock/react@2.1.0) (2020-02-12)
+
+
+### Features
+
+* expose useCodeblock hook, add isContainer flag ([a715ee9](https://github.com/codeblockjs/codeblock/commit/a715ee9a425deef93f6bccc05ec2212f758da463))
+
+
+
+
+
+# 2.0.0 (2020-02-12)
+
+
+### Features
+
+* expose async and empty provider configs ([eecb909](https://github.com/codeblockjs/codeblock/commit/eecb909e79dc2d199a0ddd95a306824f3af1894a))
+
+
+
+## 1.1.1 (2020-02-12)
+
+
+### Bug Fixes
+
+* include root-level modules in publishing ([ff5dae4](https://github.com/codeblockjs/codeblock/commit/ff5dae4f59e82cd89693b630d1f6f661516b41cb))
+
+
+
+# 1.1.0 (2020-02-12)
+
+
+### Features
+
+* expose hooks ([bb45ca1](https://github.com/codeblockjs/codeblock/commit/bb45ca1df8c8e75b1d7d1ca93b1c29f442a264e6))
+
+
+
+## 1.0.1 (2020-02-12)
+
+
+### Bug Fixes
+
+* **autoloader:** do not unset the autoloader path when unmounting a http-component ([04dbc58](https://github.com/codeblockjs/codeblock/commit/04dbc5806c24fa58ac19e4818f4aa8e9669175ef))
+
+
+
+# 1.0.0 (2020-02-12)
+
+
+### Features
+
+* add default exports for React.lazy usage ([a226ac1](https://github.com/codeblockjs/codeblock/commit/a226ac11338f407bc8f9035db109d55c25640227))
+* add http and cdn providers ([b13b907](https://github.com/codeblockjs/codeblock/commit/b13b9076ca2a0ddf637bc2e102da6490f6b66a2e))
+* add http language provider ([72bf8c2](https://github.com/codeblockjs/codeblock/commit/72bf8c2a2db8557a65fc39e009cd420561b507a2))
+
+
+### BREAKING CHANGES
+
+* drop @codeblock/styled
+
+
+
+## 0.0.2 (2020-02-10)
+
+
+
+
+
+## [1.1.1](https://github.com/codeblockjs/codeblock/compare/v1.1.0...v1.1.1) (2020-02-12)
+
+
+### Bug Fixes
+
+* include root-level modules in publishing ([ff5dae4](https://github.com/codeblockjs/codeblock/commit/ff5dae4f59e82cd89693b630d1f6f661516b41cb))
+
+
+
+
+
+# [1.1.0](https://github.com/codeblockjs/codeblock/compare/v1.0.1...v1.1.0) (2020-02-12)
+
+
+### Features
+
+* expose hooks ([bb45ca1](https://github.com/codeblockjs/codeblock/commit/bb45ca1df8c8e75b1d7d1ca93b1c29f442a264e6))
+
+
+
+
+
 ## [1.0.1](https://github.com/codeblockjs/codeblock/compare/v1.0.0...v1.0.1) (2020-02-12)