+++ packages/codeblock-themes/CHANGELOG.md
@@ -3,6 +3,103 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.0.7](https://github.com/codeblockjs/codeblock/compare/@codeblock/themes@2.0.7-alpha.0...@codeblock/themes@2.0.7) (2020-02-14)
+
+**Note:** Version bump only for package @codeblock/themes
+
+
+
+
+
+## [2.0.7-alpha.0](https://github.com/codeblockjs/codeblock/compare/@codeblock/themes@2.0.6...@codeblock/themes@2.0.7-alpha.0) (2020-02-14)
+
+
+### Features
+
+* use core to create http and cdn  providers ([720ef29](https://github.com/codeblockjs/codeblock/commit/720ef29a961d70b565a4801a2430989510aaaf17))
+
+
+
+
+
+## [2.0.6](https://github.com/codeblockjs/codeblock/compare/@codeblock/themes@2.0.5...@codeblock/themes@2.0.6) (2020-02-13)
+
+**Note:** Version bump only for package @codeblock/themes
+
+
+
+
+
+## [2.0.5](https://github.com/codeblockjs/codeblock/compare/@codeblock/themes@2.0.4...@codeblock/themes@2.0.5) (2020-02-13)
+
+**Note:** Version bump only for package @codeblock/themes
+
+
+
+
+
+## [2.0.4](https://github.com/codeblockjs/codeblock/compare/@codeblock/themes@2.0.3...@codeblock/themes@2.0.4) (2020-02-13)
+
+
+### Bug Fixes
+
+* **utils:** do not break URLs when replacing double slashes ([88bb398](https://github.com/codeblockjs/codeblock/commit/88bb398303e86d3ff619c000ee2191ed39ed6477))
+
+
+
+
+
+## [2.0.3](https://github.com/codeblockjs/codeblock/compare/@codeblock/themes@2.0.2...@codeblock/themes@2.0.3) (2020-02-13)
+
+**Note:** Version bump only for package @codeblock/themes
+
+
+
+
+
+## [2.0.2](https://github.com/codeblockjs/codeblock/compare/@codeblock/themes@2.0.1...@codeblock/themes@2.0.2) (2020-02-13)
+
+**Note:** Version bump only for package @codeblock/themes
+
+
+
+
+
+## [2.0.1](https://github.com/codeblockjs/codeblock/compare/@codeblock/themes@2.0.0...@codeblock/themes@2.0.1) (2020-02-12)
+
+
+### Bug Fixes
+
+* **funky:** Repair broken background image after prefixing, workaround for https://github.com/jotform/css.js/issues/28 ([3f8597f](https://github.com/codeblockjs/codeblock/commit/3f8597f8a10207e75243294434d37418dbd278b8))
+
+
+
+
+
+# 2.0.0 (2020-02-12)
+
+
+
+## 1.0.1 (2020-02-12)
+
+
+
+# 1.0.0 (2020-02-12)
+
+
+### Features
+
+* add http and cdn providers ([b13b907](https://github.com/codeblockjs/codeblock/commit/b13b9076ca2a0ddf637bc2e102da6490f6b66a2e))
+
+
+### BREAKING CHANGES
+
+* drop @codeblock/styled
+
+
+
+
+
 ## [1.0.1](https://github.com/codeblockjs/codeblock/compare/v1.0.0...v1.0.1) (2020-02-12)
 
 **Note:** Version bump only for package @codeblock/themes